backend/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-slugify
email-validator
python-multipart
pymysql

